+++ src/main/java/tv/alterNERD/VaultModTweaks/integration/mixin/IMixinVaultAltarIngredientsConfig.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright 2023 alterNERDtive.
+ * 
+ * This file is part of Vault Mod Tweaks.
+ * 
+ * Vault Mod Tweaks is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Vault Mod Tweaks is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * You should have received a copy of the GNU General Public License
+ * along with Vault Mod Tweaks.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package tv.alterNERD.VaultModTweaks.integration.mixin;
+
+import java.util.Map;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import iskallia.vault.config.altar.VaultAltarIngredientsConfig;
+import iskallia.vault.config.altar.entry.AltarIngredientEntry;
+import iskallia.vault.config.entry.LevelEntryMap;
+import iskallia.vault.util.data.WeightedList;
+
+/**
+ * Changes the {@link iskallia.vault.config.altar.VaultAltarIngredientsConfig}
+ * class for managing altar recipes.
+ */
+@Mixin(VaultAltarIngredientsConfig.class)
+public interface IMixinVaultAltarIngredientsConfig {
+    /**
+     * Makes the private level/entry Map available.
+     * @return The level/entry Map
+     */
+    @Accessor(remap = false)
+    LevelEntryMap<Map<String, WeightedList<AltarIngredientEntry>>> getLEVELS();
+}